**What this PR does**:
This PR contains a collection of TraceQL perf improvements that primarily look for opportunities at the iterator level. It also does some mild cleanup in block_traceql and extends the Benchmark to contain two new tests that nicely show some of the improvements:
```
{ resource.host.name = `foo` || resource.host.name = `bar` || resource.host.name = `baz` }
{ resource.host.name = `foo` } && { resource.host.name = `baz` }
```

Improvements:
- [Drop groups at the attribute level early](https://github.com/grafana/tempo/pull/2897/commits/259e6702265cfe95ebacfc04781122a9ce8106e3)
- [Unroll TruncateRowNumber](https://github.com/grafana/tempo/pull/2897/commits/bf4cf440d50321f67ad25c451cf2449b2f79230d)
- [Reduce unnecessary iteration. Set a "never iterated marker" so nil always means "iteration complete"](https://github.com/grafana/tempo/pull/2897/commits/bf4cf440d50321f67ad25c451cf2449b2f79230d)
- [Prevent double seeks](https://github.com/grafana/tempo/pull/2897/commits/ae818b228d6eb691ac490c592639ce3a83ef90e1)
  - This is the one I'm most concerned about. It was added to prevent a double seek to the same row numbers that occurs in the JoinIterators between `collect` and a `seekAll` [here](https://github.com/grafana/tempo/blob/main/pkg/parquetquery/iters.go#L1256-L1273). I'm wondering if I need store the definition level as well as the "lastRowNumber". I could also move the improvement to specifically guard against the double seek highlighted in the link.
- [Create and unroll EqualRowNumbers](https://github.com/grafana/tempo/pull/2897/commits/1280d11425b621cff8dab3270486709f86baa992)