@joe-elliott 2023-09-06
Benchmarks against main:
```
name                                                old time/op    new time/op     delta
BackendBlockTraceQL/spanAttNameNoMatch-8              6.08ms ± 1%     6.01ms ± 1%      ~     (p=0.056 n=5+5)
BackendBlockTraceQL/spanAttValNoMatch-8               7.56ms ±10%     7.00ms ± 2%      ~     (p=0.095 n=5+5)
BackendBlockTraceQL/spanAttValMatch-8                 89.7ms ± 2%     78.7ms ± 1%   -12.27%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-8         7.93ms ±13%     6.44ms ± 9%   -18.73%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch-8           53.9ms ±10%     47.8ms ± 1%   -11.33%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch-8    7.34ms ± 1%     6.25ms ± 2%   -14.80%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch-8      75.3ms ± 2%     70.1ms ± 2%    -6.98%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch-8          7.13ms ± 3%     6.10ms ± 1%   -14.44%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttValNoMatch-8           15.1ms ± 4%     12.9ms ± 1%   -14.94%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttValMatch-8             64.1ms ± 2%     61.2ms ±14%      ~     (p=0.151 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-8     7.60ms ± 7%     6.64ms ± 8%   -12.54%  (p=0.016 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch-8       33.7ms ±10%     27.6ms ± 2%   -18.16%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttOrs-8                   602ms ± 2%       70ms ± 2%   -88.37%  (p=0.008 n=5+5)
BackendBlockTraceQL/spansetAnd-8                       589ms ± 3%       69ms ± 3%   -88.37%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch-8                 457ms ± 2%      336ms ± 2%   -26.45%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValNoMatch-8                  385ms ± 4%      385ms ± 1%      ~     (p=0.310 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchAnd-8           7.69ms ± 3%     7.13ms ± 0%    -7.27%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr-8             388ms ± 1%      346ms ± 1%   -10.83%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValBothMatch-8               7.58ms ± 1%     7.13ms ± 2%    -5.85%  (p=0.008 n=5+5)

name                                                old speed      new speed       delta
BackendBlockTraceQL/spanAttNameNoMatch-8             278MB/s ± 1%    282MB/s ± 1%      ~     (p=0.056 n=5+5)
BackendBlockTraceQL/spanAttValNoMatch-8              365MB/s ±10%    392MB/s ± 2%      ~     (p=0.095 n=5+5)
BackendBlockTraceQL/spanAttValMatch-8                212MB/s ± 2%    242MB/s ± 1%   +13.97%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-8        216MB/s ±12%    264MB/s ± 9%   +22.28%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch-8          323MB/s ± 9%    364MB/s ± 1%   +12.51%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch-8   231MB/s ± 1%    271MB/s ± 2%   +17.39%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch-8     231MB/s ± 2%    248MB/s ± 3%    +7.52%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch-8         163MB/s ± 3%    190MB/s ± 1%   +16.85%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttValNoMatch-8          146MB/s ± 4%    172MB/s ± 1%   +17.49%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttValMatch-8            254MB/s ± 2%    267MB/s ±13%      ~     (p=0.151 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-8   95.2MB/s ± 8%  109.0MB/s ± 7%   +14.43%  (p=0.016 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch-8      502MB/s ± 9%    611MB/s ± 2%   +21.84%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttOrs-8                4.68MB/s ± 2%  31.56MB/s ± 2%  +574.89%  (p=0.008 n=5+5)
BackendBlockTraceQL/spansetAnd-8                    4.77MB/s ± 3%  32.23MB/s ± 3%  +575.64%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch-8              10.8MB/s ± 2%   14.6MB/s ± 2%   +35.99%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValNoMatch-8               13.2MB/s ± 3%   13.2MB/s ± 1%      ~     (p=0.341 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchAnd-8          151MB/s ± 3%    163MB/s ± 0%    +7.82%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr-8          50.3MB/s ± 1%   56.5MB/s ± 1%   +12.15%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValBothMatch-8             95.2MB/s ± 1%  101.1MB/s ± 2%    +6.22%  (p=0.008 n=5+5)

name                                                old MB_io/op   new MB_io/op    delta
BackendBlockTraceQL/spanAttNameNoMatch-8                1.69 ± 0%       1.69 ± 0%      ~     (all equal)
BackendBlockTraceQL/spanAttValNoMatch-8                 2.74 ± 0%       2.74 ± 0%      ~     (all equal)
BackendBlockTraceQL/spanAttValMatch-8                   19.0 ± 0%       19.0 ± 0%      ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-8           1.69 ± 0%       1.69 ± 0%      ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicMatch-8             17.4 ± 0%       17.4 ± 0%      ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch-8      1.69 ± 0%       1.69 ± 0%      ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch-8        17.4 ± 0%       17.4 ± 0%      ~     (all equal)
BackendBlockTraceQL/resourceAttNameNoMatch-8            1.16 ± 0%       1.16 ± 0%      ~     (all equal)
BackendBlockTraceQL/resourceAttValNoMatch-8             2.21 ± 0%       2.21 ± 0%      ~     (all equal)
BackendBlockTraceQL/resourceAttValMatch-8               16.3 ± 0%       16.3 ± 0%      ~     (all equal)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-8       0.72 ± 0%       0.72 ± 0%      ~     (all equal)
BackendBlockTraceQL/resourceAttIntrinsicMatch-8         16.9 ± 0%       16.9 ± 0%      ~     (all equal)
BackendBlockTraceQL/resourceAttOrs-8                    2.81 ± 0%       2.21 ± 0%   -21.48%  (p=0.008 n=5+5)
BackendBlockTraceQL/spansetAnd-8                        2.81 ± 0%       2.21 ± 0%   -21.48%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch-8                  4.91 ± 0%       4.91 ± 0%      ~     (all equal)
BackendBlockTraceQL/mixedValNoMatch-8                   5.08 ± 0%       5.08 ± 0%      ~     (all equal)
BackendBlockTraceQL/mixedValMixedMatchAnd-8             1.16 ± 0%       1.16 ± 0%      ~     (all equal)
BackendBlockTraceQL/mixedValMixedMatchOr-8              19.5 ± 0%       19.5 ± 0%      ~     (all equal)
BackendBlockTraceQL/mixedValBothMatch-8                 0.72 ± 0%       0.72 ± 0%      ~     (all equal)

name                                                old alloc/op   new alloc/op    delta
BackendBlockTraceQL/spanAttNameNoMatch-8              2.97MB ± 1%     2.93MB ± 3%      ~     (p=0.151 n=5+5)
BackendBlockTraceQL/spanAttValNoMatch-8               4.02MB ± 1%     4.01MB ± 0%      ~     (p=0.222 n=5+5)
BackendBlockTraceQL/spanAttValMatch-8                 4.96MB ± 8%     4.77MB ± 6%      ~     (p=0.421 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-8         3.00MB ± 2%     2.97MB ± 3%      ~     (p=1.000 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch-8           3.12MB ± 4%     3.19MB ± 8%      ~     (p=0.690 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch-8    3.02MB ± 2%     3.02MB ± 1%      ~     (p=0.690 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch-8      3.52MB ±13%     3.34MB ±12%      ~     (p=0.548 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch-8          2.95MB ± 3%     2.91MB ± 2%      ~     (p=0.548 n=5+5)
BackendBlockTraceQL/resourceAttValNoMatch-8           5.51MB ± 3%     5.47MB ± 3%      ~     (p=0.690 n=5+5)
BackendBlockTraceQL/resourceAttValMatch-8             5.65MB ± 4%     5.68MB ± 7%      ~     (p=1.000 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-8     2.93MB ± 2%     2.95MB ± 2%      ~     (p=0.421 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch-8       3.17MB ± 8%     3.20MB ± 4%      ~     (p=1.000 n=5+5)
BackendBlockTraceQL/resourceAttOrs-8                  29.1MB ± 2%     12.1MB ± 2%   -58.44%  (p=0.008 n=5+5)
BackendBlockTraceQL/spansetAnd-8                      28.7MB ± 2%     12.0MB ± 5%   -58.03%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch-8                6.80MB ±17%     6.82MB ±11%      ~     (p=0.841 n=5+5)
BackendBlockTraceQL/mixedValNoMatch-8                 7.60MB ±17%     7.68MB ±19%      ~     (p=1.000 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchAnd-8           2.98MB ± 2%     2.92MB ± 4%      ~     (p=0.095 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr-8            6.20MB ±12%     6.41MB ±15%      ~     (p=0.841 n=5+5)
BackendBlockTraceQL/mixedValBothMatch-8               2.99MB ± 2%     2.93MB ± 4%      ~     (p=0.222 n=5+5)

name                                                old allocs/op  new allocs/op   delta
BackendBlockTraceQL/spanAttNameNoMatch-8               44.0k ± 0%      44.0k ± 0%    -0.00%  (p=0.016 n=5+4)
BackendBlockTraceQL/spanAttValNoMatch-8                44.1k ± 0%      44.1k ± 0%      ~     (all equal)
BackendBlockTraceQL/spanAttValMatch-8                  53.4k ± 0%      53.4k ± 0%      ~     (p=0.294 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-8          44.0k ± 0%      44.0k ± 0%      ~     (p=0.484 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch-8            45.3k ± 0%      45.3k ± 0%      ~     (p=0.603 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch-8     44.4k ± 0%      44.4k ± 0%      ~     (p=1.000 n=5+4)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch-8       45.9k ± 0%      45.9k ± 0%      ~     (p=0.381 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch-8           44.0k ± 0%      44.0k ± 0%      ~     (p=0.651 n=5+4)
BackendBlockTraceQL/resourceAttValNoMatch-8            44.2k ± 0%      44.2k ± 0%      ~     (p=0.603 n=5+5)
BackendBlockTraceQL/resourceAttValMatch-8              45.5k ± 0%      45.5k ± 0%      ~     (p=0.421 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-8      44.0k ± 0%      44.0k ± 0%      ~     (p=1.000 n=5+4)
BackendBlockTraceQL/resourceAttIntrinsicMatch-8        45.4k ± 0%      45.4k ± 0%      ~     (p=0.587 n=5+5)
BackendBlockTraceQL/resourceAttOrs-8                    624k ± 0%       431k ± 0%   -31.02%  (p=0.008 n=5+5)
BackendBlockTraceQL/spansetAnd-8                        624k ± 0%       431k ± 0%   -31.02%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch-8                 44.6k ± 0%      44.6k ± 0%      ~     (p=0.444 n=5+5)
BackendBlockTraceQL/mixedValNoMatch-8                  48.2k ± 0%      47.1k ± 0%    -2.31%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchAnd-8            44.0k ± 0%      44.0k ± 0%      ~     (p=0.135 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr-8             45.8k ± 0%      45.8k ± 0%      ~     (p=0.802 n=5+5)
BackendBlockTraceQL/mixedValBothMatch-8                44.0k ± 0%      44.0k ± 0%      ~     (p=0.175 n=5+5)
```